requirements.txt
attrs>=22.1.0
cachetools>=5.2.0
certifi>=2022.9.24
cffi>=1.15.1
charset-normalizer>=2.1.1
click>=8.1.3
colorama>=0.4.5
coverage>=7.5.0
cryptography>=38.0.1
defer>=1.0.4
flake8>=5.0.4
Flask>=2.2.2
Flask-Login>=0.6.2
Flask-SQLAlchemy>=2.5.1
google-auth-oauthlib>=1.0.0
google-api-core>=2.10.1
google-api-python-client>=2.48.0
google-auth>=2.15.0
google-auth-httplib2>=0.1.0
greenlet>=1.1.3
httplib2>=0.20.4
idna>=3.4
importlib-metadata>=5.0.0
iniconfig>=1.1.1
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.1
packaging>=21.3
Pillow>=9.2.0
pluggy>=1.0.0
protobuf>=4.21.7
py>=1.11.0
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycparser>=2.21
pyfiglet>=0.8.post1
pyflakes>=2.5.0
PyMySQL>=1.0.2
pyparsing>=3.0.9
pytest>=7.1.3
python-resize-image>=1.1.20
requests>=2.28.1
rsa>=4.9
six>=1.16.0
SQLAlchemy>=1.4.41
termcolor>=1.1.0
tomli>=2.0.1
uritemplate>=4.1.1
urllib3>=1.26.12
Werkzeug>=2.2.2
zipp>=3.8.1
Google-Images-Search>=1.4.7
google-generativeai>=0.8.3
geocoder>=1.38.1
google-genai
beautifulsoup4>=4.13.3